pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autostrategy"
version = "0.1.0"
description = "Autoresearch for trading strategies - AI writes, tests, and evolves trading strategies overnight"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "The Menon Lab", email = "hello@themenonlab.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.18.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.36",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
autostrategy = "autostrategy.main:main"

[project.urls]
Homepage = "https://github.com/menonpg/autostrategy"
Repository = "https://github.com/menonpg/autostrategy"
Documentation = "https://github.com/menonpg/autostrategy#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["autostrategy*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
